# backend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Forum.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5032/"
    }
  ],
  "paths": {
...
    "/api/v1/posts": {
      "post": {
        "tags": [
          "Posts"
        ],
        "summary": "Create a post.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/posts/{id}": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Read a single post with its first page of comments.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePostRequest": {
        "required": [
          "title",
          "body"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}